requirements.txt
aiohttp==3.8.6
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
blessed==1.20.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorful==0.5.5
decorator==5.1.1
diskcache==5.6.3
distlib==0.3.7
executing==2.0.1
fastapi==0.104.1
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.10.0
google-api-core==2.12.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
gpustat==1.1.1
grpcio==1.59.2
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.17.3
idna==3.4
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
llama_cpp_python==0.2.15
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.2.1
msgpack==1.0.7
multidict==6.0.4
networkx==3.0rc1
nltk==3.8.1
numpy==1.24.1
nvidia-ml-py==12.535.133
opencensus==0.11.3
opencensus-context==0.1.3
packaging==23.2
parso==0.8.3
pexpect==4.8.0
Pillow==9.3.0
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
python-dotenv==1.0.0
PyYAML==6.0.1
ray==2.8.0
referencing==0.30.2
regex==2023.10.3
requests==2.28.1
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sse-starlette==1.6.5
stack-data==0.6.3
starlette==0.27.0
starlette-context==0.3.6
sympy==1.11.1
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.2.0.dev20231107
torchaudio==2.2.0.dev20231107
torchvision==0.17.0.dev20231107
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
typing_extensions==4.8.0
urllib3==1.26.13
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.21.0
watchfiles==0.21.0
wcwidth==0.2.9
websockets==12.0
yarl==1.9.2
